# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OctoManger API v2
  version: 2.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v2/email/accounts:
    get:
      operationId: listEmailAccounts
      responses:
        "200":
          description: Email accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/EmailAccount"
    post:
      operationId: createEmailAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailAccountCreateInput"
      responses:
        "201":
          description: Created email account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailAccount"
  /api/v2/email/accounts/bulk-import:
    post:
      operationId: bulkImportEmailAccounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailBulkImportInput"
      responses:
        "200":
          description: Bulk import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailBulkImportResult"
  /api/v2/email/accounts/{id}:
    get:
      operationId: getEmailAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email account detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailAccount"
    patch:
      operationId: patchEmailAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailAccountPatchInput"
      responses:
        "200":
          description: Updated email account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailAccount"
    delete:
      operationId: deleteEmailAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted email account
  /api/v2/email/accounts/{id}/outlook/authorize-url:
    post:
      operationId: buildOutlookAuthorizeURL
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Outlook authorize URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutlookAuthorizeURLResult"
  /api/v2/email/accounts/{id}/outlook/exchange-code:
    post:
      operationId: exchangeOutlookCode
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutlookExchangeCodeInput"
      responses:
        "200":
          description: Email account updated with exchanged Outlook tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailAccount"
  /api/v2/email/accounts/{id}/mailboxes:
    get:
      operationId: listEmailMailboxes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: pattern
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Mailboxes available for the configured Outlook account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailMailboxListResult"
  /api/v2/email/accounts/{id}/messages:
    get:
      operationId: listEmailMessages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: mailbox
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Messages in a mailbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailMessageListResult"
  /api/v2/email/accounts/{id}/messages/latest:
    get:
      operationId: getLatestEmailMessage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: mailbox
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Latest message in the requested mailbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailLatestMessageResult"
  /api/v2/email/accounts/{id}/messages/{message_id}:
    get:
      operationId: getEmailMessage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: message_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Detailed email message payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailMessageDetail"
  /api/v2/email/preview/mailboxes:
    post:
      operationId: previewEmailMailboxes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailPreviewInput"
      responses:
        "200":
          description: Mailbox preview using a draft Outlook config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailMailboxListResult"
  /api/v2/email/preview/messages/latest:
    post:
      operationId: previewLatestEmailMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailPreviewInput"
      responses:
        "200":
          description: Latest message preview using a draft Outlook config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailLatestMessageResult"
...
components:
  schemas:
...
    EmailAccount:
      type: object
      required:
        - id
        - address
        - provider
        - status
        - config
      properties:
        id:
          type: integer
        address:
          type: string
        provider:
          type: string
        status:
          type: string
        config:
          type: object
          additionalProperties: true
    EmailAccountCreateInput:
      type: object
      required:
        - address
        - provider
        - status
        - config
      properties:
        address:
          type: string
        provider:
          type: string
        status:
          type: string
        config:
          type: object
          additionalProperties: true
    EmailAccountPatchInput:
      type: object
      properties:
        provider:
          type: string
        status:
          type: string
        config:
          type: object
          additionalProperties: true
    EmailMailbox:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
    EmailMailboxListResult:
      type: object
      required:
        - items
      properties:
        pattern:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/EmailMailbox"
    EmailMessageSummary:
      type: object
      required:
        - id
        - subject
        - from
        - to
        - date
        - size
        - flags
      properties:
        id:
          type: string
        subject:
          type: string
        from:
          type: string
        to:
          type: string
        date:
          type: string
          format: date-time
        size:
          type: integer
        flags:
          type: array
          items:
            type: string
    EmailMessageListResult:
      type: object
      required:
        - mailbox
        - limit
        - offset
        - total
        - items
      properties:
        mailbox:
          type: string
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/EmailMessageSummary"
    EmailMessageDetail:
      type: object
      required:
        - id
        - subject
        - from
        - to
        - cc
        - date
        - size
        - flags
        - headers
        - text_body
        - html_body
      properties:
        id:
          type: string
        subject:
          type: string
        from:
          type: string
        to:
          type: string
        cc:
          type: string
        date:
          type: string
          format: date-time
        size:
          type: integer
        flags:
          type: array
          items:
            type: string
        headers:
          type: object
          additionalProperties:
            type: string
        text_body:
          type: string
        html_body:
          type: string
    EmailLatestMessageResult:
      type: object
      required:
        - mailbox
        - found
      properties:
        mailbox:
          type: string
        found:
          type: boolean
        item:
          $ref: "#/components/schemas/EmailMessageDetail"
    OutlookAuthorizeURLResult:
      type: object
      required:
        - authorize_url
      properties:
        authorize_url:
          type: string
          format: uri
    OutlookExchangeCodeInput:
      type: object
      properties:
        code:
          type: string
      required:
        - code
    EmailPreviewInput:
      type: object
      properties:
        config:
          type: object
          additionalProperties: true
        mailbox:
          type: string
        pattern:
          type: string
      required:
        - config
    EmailBulkImportInput:
      type: object
      required:
        - lines
      properties:
        lines:
          type: array
          items:
            type: string
    EmailBulkImportLineResult:
      type: object
      required:
        - line
        - ok
      properties:
        line:
          type: string
        address:
          type: string
        ok:
          type: boolean
        error:
          type: string
        id:
          type: integer
    EmailBulkImportResult:
      type: object
      required:
        - total
        - success
        - failed
        - items
      properties:
        total:
          type: integer
        success:
          type: integer
        failed:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/EmailBulkImportLineResult"
...